Fix: chr check bug, add: count reads/bin
Added functionality of add_reads_per_bin, to which one parses the reads dataframe (i.e. the data) and a binned genome. Each bin will be updated to contain the count of reads falling within it, to allow quick shuffling followed by updating of SV per bin using a loop.